third_party/blink/web_tests/wpt_internal/html/canvas/drawElementImage/privacy/box-mask-png-ignored.https.sub.html

<!DOCTYPE HTML>
<html>
<head>
  <title>drawElementImage does not use cross-origin mask images</title>
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
  <style>
  #child {
    width: 100px;
    height: 200px;
    background: green;
  }
  #sameOrigin {
    position: absolute;
    left: 0px;
    top: 0px;
    width: 100px;
    height: 100px;
    background-color: red;
    mask-image: url("https://{{location[host]}}/wpt_internal/html/canvas/drawElementImage/resources/blue-100x50-transparent-100x50.png");
  }
  #crossOrigin {
    position: absolute;
    left: 0px;
    top: 100px;
    width: 100px;
    height: 100px;
    background-color: red;
    mask-image: url("https://{{hosts[alt][www]}}:{{ports[h2][0]}}/wpt_internal/html/canvas/drawElementImage/resources/blue-100x50-transparent-100x50.png");
  }
  </style>
</head>
<body>
  <canvas id=canvas width="100" height="200" layoutsubtree>
    <div id=child>
      <div id=sameOrigin></div>
      <div id=crossOrigin></div>
    </div>
  </canvas>

  <script>
    window.onload = () => {
      test(function(t) {
        var context = canvas.getContext("2d");
        context.drawElementImage(child, 0, 0);

        pixel = context.getImageData(50, 25, 1, 1).data;
        assert_equals(pixel[0], 0, "Same origin should draw");
        assert_equals(pixel[1], 128, "Same origin should draw");
        assert_equals(pixel[2], 0, "Same origin should draw");

        pixel = context.getImageData(50, 75, 1, 1).data;
        assert_equals(pixel[0], 255, "Same origin should draw");
        assert_equals(pixel[1], 0, "Same origin should draw");
        assert_equals(pixel[2], 0, "Same origin should draw");

        // Missing mask image renders as black, blocking the red background
        pixel = context.getImageData(50, 125, 1, 1).data;
        assert_equals(pixel[0], 0, "Cross origin should not draw");
        assert_equals(pixel[1], 128, "Cross origin should not draw");
        assert_equals(pixel[2], 0, "Cross origin should not draw");

        pixel = context.getImageData(50, 175, 1, 1).data;
        assert_equals(pixel[0], 0, "Cross origin should not draw");
        assert_equals(pixel[1], 128, "Cross origin should not draw");
        assert_equals(pixel[2], 0, "Cross origin should not draw");

        t.done();
      });
    }
  </script>
</body>
</html>
